+++ src/main/java/executor/service/model/config/WebDriverConfig.java
@@ -1,4 +1,4 @@
-package executor.service.model;
+package executor.service.model.config;
 
 import java.util.Objects;
