+++ src/main/java/executor/service/model/scenario/Scenario.java
@@ -1,4 +1,4 @@
-package executor.service.model;
+package executor.service.model.scenario;
 
 import java.util.Objects;
 import java.util.List;
