+++ src/main/java/executor/service/model/scenario/Step.java
@@ -1,4 +1,4 @@
-package executor.service.model;
+package executor.service.model.scenario;
 
 import java.util.Objects;
