+++ src/main/java/executor/service/service/proxy/ProxySourcesClientImpl.java
@@ -1,7 +1,6 @@
-package executor.service.service.impl;
+package executor.service.service.proxy;
 
-import executor.service.model.ProxyConfigHolder;
-import executor.service.service.ProxySourcesClient;
+import executor.service.model.proxy.ProxyConfigHolder;
 
 public class ProxySourcesClientImpl implements ProxySourcesClient {
     @Override
