+++ src/main/java/executor/service/service/scenario/ScenarioExecutor.java
@@ -1,6 +1,6 @@
-package executor.service.service;
+package executor.service.service.scenario;
 
-import executor.service.model.Scenario;
+import executor.service.model.scenario.Scenario;
 import org.openqa.selenium.WebDriver;
 
 public interface ScenarioExecutor {
