+++ src/main/java/executor/service/service/scenario/ScenarioSourceListener.java
@@ -1,4 +1,4 @@
-package executor.service.service;
+package executor.service.service.scenario;
 
 public interface ScenarioSourceListener {
