+++ src/main/java/executor/service/service/stepexections/StepExecutionSleepImpl.java
@@ -1,7 +1,6 @@
-package executor.service.service.impl;
+package executor.service.service.stepexections;
 
-import executor.service.model.Step;
-import executor.service.service.StepExecutionSleep;
+import executor.service.model.scenario.Step;
 import org.openqa.selenium.WebDriver;
 
 public class StepExecutionSleepImpl implements StepExecutionSleep {
