+++ src/main/java/executor/service/service/webdriver/WebDriverInitializer.java
@@ -1,7 +1,7 @@
-package executor.service.service;
+package executor.service.service.webdriver;
 
-import executor.service.model.ProxyConfigHolder;
-import executor.service.model.WebDriverConfig;
+import executor.service.model.proxy.ProxyConfigHolder;
+import executor.service.model.config.WebDriverConfig;
 import org.openqa.selenium.WebDriver;
 
 public class WebDriverInitializer {
