+++ src/test/java/executor/service/model/ProxyNetworkConfigTest.java
@@ -1,5 +1,6 @@
 package executor.service.model;
 
+import executor.service.model.proxy.ProxyNetworkConfig;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
