+++ src/test/java/executor/service/model/StepTest.java
@@ -1,4 +1,5 @@
 package executor.service.model;
+import executor.service.model.scenario.Step;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
